Relaxed warning level with ffmpeg, and fixed remaining warnings. It now compiles.

+++ lib/ffmpeg/include/libavutil/common.h
@@ -136,8 +136,8 @@ static av_always_inline av_const int64_t av_clip64_c(int64_t a, int64_t amin, in
  */
 static av_always_inline av_const uint8_t av_clip_uint8_c(int a)
 {
-    if (a&(~0xFF)) return (-a)>>31;
-    else           return a;
+    if (a&(~0xFF)) return uint8_t((-a)>>31);
+    else           return uint8_t(a);
 }
 
 /**
@@ -147,8 +147,8 @@ static av_always_inline av_const uint8_t av_clip_uint8_c(int a)
  */
 static av_always_inline av_const int8_t av_clip_int8_c(int a)
 {
-    if ((a+0x80) & ~0xFF) return (a>>31) ^ 0x7F;
-    else                  return a;
+    if ((a+0x80) & ~0xFF) return uint8_t((a>>31) ^ 0x7F);
+    else                  return uint8_t(a);
 }
 
 /**
@@ -158,8 +158,8 @@ static av_always_inline av_const int8_t av_clip_int8_c(int a)
  */
 static av_always_inline av_const uint16_t av_clip_uint16_c(int a)
 {
-    if (a&(~0xFFFF)) return (-a)>>31;
-    else             return a;
+    if (a&(~0xFFFF)) return uint16_t((-a)>>31);
+    else             return uint16_t(a);
 }
 
 /**
@@ -169,8 +169,8 @@ static av_always_inline av_const uint16_t av_clip_uint16_c(int a)
  */
 static av_always_inline av_const int16_t av_clip_int16_c(int a)
 {
-    if ((a+0x8000) & ~0xFFFF) return (a>>31) ^ 0x7FFF;
-    else                      return a;
+    if ((a+0x8000) & ~0xFFFF) return int16_t((a>>31) ^ 0x7FFF);
+    else                      return int16_t(a);
 }
 
 /**
